--- PluginCaller/pch.h
@@ -1,3 +1,0 @@
-﻿#pragma once
-#include <winrt/Windows.Foundation.h>
-#include <winrt/Windows.Foundation.Collections.h>
