+++ app2app.sln
@@ -12,8 +12,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Items", "Items", "{A06277EA
 EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PluginApp1", "PluginApp1\PluginApp1.vcxproj", "{E3347C64-B8C4-4DB0-A5C1-C239C137CC7C}"
 EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PluginCaller", "PluginCaller\PluginCaller.vcxproj", "{C058FFDC-39C6-4B1F-891C-AE381227D11F}"
-EndProject
 Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "PluginCaller2", "PluginCaller2\PluginCaller2.vcxproj", "{9FAFD732-73C8-4633-B2E3-7B8D9C19F947}"
 EndProject
 Global
@@ -68,22 +66,6 @@ Global
 		{E3347C64-B8C4-4DB0-A5C1-C239C137CC7C}.Release|x86.ActiveCfg = Release|Win32
 		{E3347C64-B8C4-4DB0-A5C1-C239C137CC7C}.Release|x86.Build.0 = Release|Win32
 		{E3347C64-B8C4-4DB0-A5C1-C239C137CC7C}.Release|x86.Deploy.0 = Release|Win32
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Debug|ARM.ActiveCfg = Debug|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Debug|ARM.Build.0 = Debug|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Debug|ARM64.ActiveCfg = Debug|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Debug|ARM64.Build.0 = Debug|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Debug|x64.ActiveCfg = Debug|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Debug|x64.Build.0 = Debug|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Debug|x86.ActiveCfg = Debug|Win32
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Debug|x86.Build.0 = Debug|Win32
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Release|ARM.ActiveCfg = Release|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Release|ARM.Build.0 = Release|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Release|ARM64.ActiveCfg = Release|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Release|ARM64.Build.0 = Release|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Release|x64.ActiveCfg = Release|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Release|x64.Build.0 = Release|x64
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Release|x86.ActiveCfg = Release|Win32
-		{C058FFDC-39C6-4B1F-891C-AE381227D11F}.Release|x86.Build.0 = Release|Win32
 		{9FAFD732-73C8-4633-B2E3-7B8D9C19F947}.Debug|ARM.ActiveCfg = Debug|x64
 		{9FAFD732-73C8-4633-B2E3-7B8D9C19F947}.Debug|ARM.Build.0 = Debug|x64
 		{9FAFD732-73C8-4633-B2E3-7B8D9C19F947}.Debug|ARM.Deploy.0 = Debug|x64
